lots of reorganisation; move token.rb up a folder; re-scope validators
