+++ spec/tokenizer/lexer/comment_spec.rb
@@ -1,5 +1,5 @@
+require './src/token.rb'
 require './src/tokenizer/lexer.rb'
-require './src/tokenizer/token.rb'
 require './spec/contexts/lexer.rb'
 
 include Tokenizer
