+++ spec/tokenizer/lexer/properties_spec.rb
@@ -1,6 +1,5 @@
+require './src/token.rb'
 require './src/tokenizer/lexer.rb'
-require './src/tokenizer/token.rb'
-
 require './spec/contexts/lexer.rb'
 
 include Tokenizer
